Issue-56: Upgrade osrm to the latest 5.22.0. Refactor unit test so it can run in parallel.
