+++ libosrmver.txt
@@ -1,1 +1,1 @@
-5.18.0
+5.22.0
