+++ src/Test/osrm.net.test/osrm.net.test.csproj
@@ -82,18 +82,16 @@
   <ItemGroup>
     <Compile Include="Common\IledeFranceEngineConfigMld.cs" />
     <Compile Include="Common\OsrmTestAssert.cs" />
-    <Compile Include="Routing\WhenUsingIldeFranceMld.cs" />
-    <Compile Include="Routing\WhenUsingInvalidEngineConfig.cs" />
-    <Compile Include="Routing\WhenUsingOrlandoCh.cs" />
+    <Compile Include="Tests\WhenUsingIldeFranceMld.cs" />
+    <Compile Include="Tests\WhenUsingInvalidEngineConfig.cs" />
+    <Compile Include="Tests\WhenUsingOrlandoCh.cs" />
     <Compile Include="Common\BaseEngineConfig.cs" />
     <Compile Include="Common\InvalidEngineConfig.cs" />
     <Compile Include="Common\OrlandoEngineConfigCh.cs" />
     <Compile Include="Common\OrlandoEngineConfigMld.cs" />
     <Compile Include="EngineConfig\EngineConfigTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Routing\WhenUsingOrlandoMld.cs" />
-    <Compile Include="Table\WhenUsingOrlandoCh.cs" />
-    <Compile Include="Table\WhenUsingOrlandoMld.cs" />
+    <Compile Include="Tests\WhenUsingOrlandoMld.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
